Fix tests, except some are broken
