+++ app/components/bulletin-component.js
@@ -47,4 +47,4 @@ export default PdfItem.extend({
       }
     }
   }
-})
+});
