+++ bower.json
@@ -9,7 +9,7 @@
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
     "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
     "ember-load-initializers": "ember-cli/ember-load-initializers#0.0.2",
-    "ember-qunit": "0.3.1",
+    "ember-qunit": "0.4.x",
     "ember-qunit-notifications": "0.0.7",
     "qunit": "~1.17.1",
     "moment": "~2.9.0",
